Commit initial avec Svelte
Une reproduction du comportement de la branche `main` en utilisant `Svelte` au lieu de `React`.

`Prettier` est aussi disponible manuellement en utilisant `npm run format`.
