--- src/App.css
@@ -1,11 +1,0 @@
-#root {
-  width: 1000px;
-  padding: 2rem;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-}
-
-.card {
-  padding-top: 10px;
-}
