+++ src/vite-env.d.ts
@@ -1,1 +1,2 @@
+/// <reference types="svelte" />
 /// <reference types="vite/client" />
