+++ tsconfig.node.json
@@ -4,7 +4,6 @@
     "skipLibCheck": true,
     "module": "ESNext",
     "moduleResolution": "bundler",
-    "allowSyntheticDefaultImports": true,
     "strict": true
   },
   "include": ["vite.config.ts"]
